PhysX now works on XAGE.

+++ XPXNgin.Physics/Source/Nplicit.h
@@ -14,5 +14,5 @@
 
 namespace XPX
 {
-	typedef double NplicitFloat;
+	typedef float NplicitFloat;
 }
